data_generation/posture/requirements.txt
huggingface_hub
python-dotenv
torch
torchvision
Pillow
diffusers
controlnet_aux
accelerate
transformers
scikit-learn
onnx
